backend/requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
PyJWT 